requirements.txt
# PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# Embeddings
sentence-transformers>=2.2.0
openai>=1.0.0

# Vector Store
requests>=2.31.0

# LLM
openai>=1.0.0
google-genai>=0.1.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
tqdm>=4.66.0
colorama>=0.4.6
